+++ castle/cms/utils/content.py
@@ -14,7 +14,7 @@ from castle.cms.interfaces import IReferenceNamedImage
 from OFS.CopySupport import CopyError
 from OFS.CopySupport import _cb_decode
 from OFS.CopySupport import eInvalid
-from plone import api
+import plone.api as api
 from plone.app.blocks.layoutbehavior import ILayoutAware
 from plone.app.contentlisting.interfaces import IContentListingObject
 from plone.app.uuid.utils import uuidToCatalogBrain
@@ -87,7 +87,7 @@ def clear_object_cache(ob):
     ob._p_jar.sync()
 
 
-def inline_images_in_dom(dom, portal=None, site_url=None):
+def inline_images_in_dom(dom, portal=None, site_url=None, unrestricted_traverse=False):
     if portal is None:
         portal = api.portal.get()
     if site_url is None:
@@ -95,7 +95,12 @@ def inline_images_in_dom(dom, portal=None, site_url=None):
     for img in dom.cssselect('img'):
         src = img.attrib.get('src', '')
         if src.startswith(site_url):
-            ctype, data = get_data_from_url(src, portal, site_url)
+            ctype, data = get_data_from_url(
+                url=src,
+                portal=portal,
+                site_url=site_url,
+                unrestricted_traverse=unrestricted_traverse,
+            )
             if not ctype or not data or 'image' not in ctype:
                 img.attrib['src'] = ''
                 continue
@@ -104,11 +109,14 @@ def inline_images_in_dom(dom, portal=None, site_url=None):
             img.attrib['src'] = data_uri
 
 
-def get_data_from_url(url, portal=None, site_url=None):
+def get_data_from_url(url, portal=None, site_url=None, unrestricted_traverse=False):
     if portal is None:
         portal = api.portal.get()
     if site_url is None:
         site_url = portal.absolute_url()
+    traverse_function = portal.restrictedTraverse
+    if unrestricted_traverse:
+        traverse_function = portal.unrestrictedTraverse
     data = None
     path = url.replace(site_url, '').strip('/')
     ct = ''
@@ -119,7 +127,7 @@ def get_data_from_url(url, portal=None, site_url=None):
             if len(parts) == 2:
                 fieldname = parts[0]
                 size = parts[1]
-                images = portal.restrictedTraverse(
+                images = traverse_function(
                     str(path + '/@@images'), None)
                 if images is not None:
                     im = images.traverse(fieldname, [size])
@@ -130,14 +138,14 @@ def get_data_from_url(url, portal=None, site_url=None):
                         pass
             else:
                 # grab full size
-                ob = portal.restrictedTraverse(str(path), None)
+                ob = traverse_function(str(path), None)
                 data = ob.image.data
                 ct = ob.image.contentType
         except Exception:
             logger.error('Could not traverse image ' + url, exc_info=True)
 
     if data is None:
-        ob = portal.restrictedTraverse(str(path), None)
+        ob = traverse_function(str(path), None)
         file_path = getattr(ob, 'path', None)
         if file_path is None:
             try: